Adding level files for the Game, Level 1,2 and 3
This Commit Adds the files for the Digit Defence. 
The files are the core files for the game and contains the functions which are called by the main function. 
Comments on the files for Level2.cpp and Level3.cpp are omitted as it is the same as Level1.cpp
To see detailed comments, please open Level1.cpp.